refactor html-entrypoint to its own module
